FY-131/refactor: 검색 바 컴포넌트 분리
